requirements.txt
streamlit==0.81.1
pandas==1.3.1
plotly==4.14.3
matplotlib==3.2.1
Pillow==7.1.2





